Move inception test helpers to separate file
